requirements.txt
openai
python-dotenv
faiss-cpu
numpy
tqdm
pypdf
langchain
langchain-community